+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.Controller.race_controller;
+
 public class Application {
     public static void main(String[] args) {
         race_controller race_controller = new race_controller();
