+++ src/main/java/racingcar/Model/Cars.java
@@ -1,11 +1,10 @@
-package racingcar;
+package racingcar.Model;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
+import racingcar.constant.ErrorMessage;
 
 public class Cars {
     private final List<Car> cars;
