+++ src/main/java/racingcar/Model/Cars_preparation.java
@@ -1,13 +1,11 @@
-package racingcar;
+package racingcar.Model;
 
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
+import racingcar.Model.Car;
 
 public class Cars_preparation {
     private final Map<Car, Integer> carAndPositionMap;
