+++ src/test/java/racingcar/ErrorTest/carTest.java
@@ -1,10 +1,11 @@
-package racingcar;
+package racingcar.ErrorTest;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.Test;
+import racingcar.Application;
 
 public class carTest extends NsTest {
     @Test
